requirements.txt
Flask
flask-cors
flask-bcrypt
flask-jwt-extended
mysql-connector-python
